Added multi extension to get file

+++ DatabaseEncryption/Helper/SqlCipherHelper.cs
@@ -14,52 +14,57 @@ namespace DatabaseEncryption
 
         public void DoEncription(string path)
         {
+            string[] extensions = { ".db", ".sqlite" };
             DirectoryInfo d = new DirectoryInfo($@"{path}");
-            FileInfo[] files = d.GetFiles("*.db");
+            FileInfo[] files = d.GetFiles();
 
             foreach (FileInfo file in files)
             {
-                string enc_db_path = $@"{path}\\enc_{file.Name.Split('.')[0]}.sqlite";
+                if (extensions.Any(x => x.Equals(file.Extension, StringComparison.CurrentCultureIgnoreCase)))
+                {
+                    string enc_db_path = $@"{path}\\enc_{file.Name.Split('.')[0]}.sqlite";
 
-                if (File.Exists(enc_db_path))
-                    File.Delete(enc_db_path);
+                    if (File.Exists(enc_db_path))
+                        File.Delete(enc_db_path);
 
-                SqliteConnection sqlConnection = new SqliteConnection($@"Data Source={file.FullName};"); 
-                sqlConnection.Open();
+                    SqliteConnection sqlConnection = new SqliteConnection($@"Data Source={file.FullName};");
+                    sqlConnection.Open();
 
-                string sql = "PRAGMA cipher_version";
-                SqliteCommand command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    string sql = "PRAGMA cipher_version";
+                    SqliteCommand command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = "ATTACH DATABASE '" + enc_db_path + "' AS encrypted KEY '" + this.options.Password + "';";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = "ATTACH DATABASE '" + enc_db_path + "' AS encrypted KEY '" + this.options.Password + "';";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = $"PRAGMA encrypted.cipher_page_size = {this.options.CipherPageSize};";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = $"PRAGMA encrypted.cipher_page_size = {this.options.CipherPageSize};";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = $"PRAGMA encrypted.kdf_iter = {this.options.KdfIter};";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = $"PRAGMA encrypted.kdf_iter = {this.options.KdfIter};";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = $"PRAGMA encrypted.cipher_hmac_algorithm = {this.options.CipherHmacAlgorithm};";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = $"PRAGMA encrypted.cipher_hmac_algorithm = {this.options.CipherHmacAlgorithm};";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = $"PRAGMA encrypted.cipher_kdf_algorithm  = {this.options.CipherKdfAlgorithm};";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = $"PRAGMA encrypted.cipher_kdf_algorithm  = {this.options.CipherKdfAlgorithm};";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = "SELECT sqlcipher_export('encrypted')";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = "SELECT sqlcipher_export('encrypted')";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sql = "DETACH DATABASE encrypted;";
-                command = new SqliteCommand(sql, sqlConnection);
-                command.ExecuteNonQuery();
+                    sql = "DETACH DATABASE encrypted;";
+                    command = new SqliteCommand(sql, sqlConnection);
+                    command.ExecuteNonQuery();
 
-                sqlConnection.Close();
+                    sqlConnection.Close();
+
+                }
 
             }
         }
